+++ lib/decorators/custom/ChyanApplication.ts
@@ -1,9 +1,9 @@
 import { initializer } from "../../core/initializer";
-import { Injectable } from "./Injectable";
+import { Injectable } from "../internal/Injectable";
 
 function init() {
   setImmediate(initializer);
-  return { isApplication: true };
+  return { isApplication: true, prefix: "" };
 }
 
 export function ChyanApplication() {